replacing shim file symlink by a shim copy new lightweight shim that only calls volta-runner
