+++ crates/volta-core/src/run/mod.rs
@@ -104,6 +104,8 @@ fn get_executor(
 
 /// Determine the name of the command to run by inspecting the first argument to the active process
 fn get_tool_name(args: &mut ArgsOs) -> Fallible<OsString> {
+    args.next();
+
     args.next()
         .and_then(|arg0| Path::new(&arg0).file_name().map(tool_name_from_file_name))
         .ok_or_else(|| ErrorKind::CouldNotDetermineTool.into())
